requirements.txt
pandas
matplotlib
PyPDF2
python-dotenv
requests
openpyxl
numpy
